updates README to new php-amqplib location
I moved the development of php-amqplib from the
tnc account to the videlalvaro account and I started
adding tests and benchmarks there.

More info here: http://videlalvaro.github.com/2011/11/php-amqplib-future.html

Now the version used for the tutorials is tagged. Once I'm done with the refactorings that I want to apply I will test the code again and and resubmit a
patch for the README