Update Ruby test suites
Testing on Ruby 1.8 is no longer a good idea: 1.8 is no longer
supported in any way, including security patches, so it will be
removed from recent OS distributions fairly soon and fewer
and fewer projects will be compatible with it.

Ruby clients run tests against 1.8.7 on travis-ci.org
and will do so for as long as it is available.