add Elixir tutorial 1
Prior to this commit, the RabbitMQ tutorials contained no Elixir
examples. One _could_ look at the Erlang examples and adapt them, but
that certainly does not result in idiomatic Elixir. This commit adds the
first tutorial and an Elixir project file specifying the `amqp` library
as the dependency for communicating with RabbitMQ via idiomatic Elixir.