Avoid ArrayIndexOutOfBoundsException in EmitLogDirect
It is true that the EmitLogDirect.getMessage() method only calls the
EmitLogDirect.joinStrings() method when the length of the strings array
argument is at least 2. However, it is safe make our method more robust
by avoiding the ArrayIndexOutOfBoundsException exception as much as we
can. Who knows? Maybe the EmitLogDirect.joinStrings() method could be
extracted in a utility class and used in cases where the startIndex
could potentially be equal to the length of the passed array.